fix: remove "customtab" from scenarios path
